requirements.txt
attrs==21.4.0
bleach==4.1.0
bokeh==1.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
colorcet==3.0.0
cycler==0.11.0
dataclasses==0.8
holoviews==1.12.7
hvplot==0.4.0
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.0.0
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
numpy==1.19.5
onnx==1.8.1
osqp==0.6.2.post5
packaging==21.3
pandas==1.1.5
param==1.12.2
Pillow==8.4.0
progressbar2==3.55.0
protobuf==3.19.4
pybind11==2.10.0
pyct==0.4.8
pyparsing==3.0.9
pyrsistent==0.18.0
python-dateutil==2.8.2
python-utils==3.3.3
pytz==2022.1
pyviz-comms==2.2.0
PyYAML==6.0
qdldl==0.1.5.post2
requests==2.27.1
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
tensorboardX==2.4
threadpoolctl==3.1.0
timm==0.6.5
torch==1.9.1+cpu
torchaudio==0.9.1
torchvision==0.10.1+cpu
tornado==6.1
tqdm==4.64.0
typing_extensions==4.1.1
urllib3==1.26.10
webencodings==0.5.1
zipp==3.6.0
